+++ api/controllers/timesheet/add.js
@@ -1,4 +1,4 @@
-/* global sails, Directors, Timesheet, moment */
+/* global sails, Directors, Timesheet, moment, Directorhours */
 
 module.exports = {
 
@@ -52,12 +52,20 @@ module.exports = {
                 var toapprove = false;
                 thetime = moment(inputs.timestamp);
 
+                // Check if an office hours record exists. Allow 30 minutes grace.
+                var calendar = await Directorhours.find({director: record.name, active: 1, start: {"<=": moment().add(30, 'minutes').toISOString(true)}, end: {">=": moment().toISOString(true)}}).limit(1);
+
                 // If the entry is less than 30 minutes off from the current time, approve automatically
                 if (thetime.isAfter(moment().subtract(30, 'minutes')) && thetime.isBefore(moment().add(30, 'minutes')))
                     toapprove = true;
 
                 // Clock-ins need a new entry
-                await Timesheet.create({name: record.name, time_in: thetime.toISOString(true), approved: toapprove}).fetch();
+                if (calendar.length > 0)
+                {
+                    await Timesheet.create({name: record.name, unique: calendar[0].unique, scheduled_in: moment(calendar[0].start).toISOString(true), scheduled_out: moment(calendar[0].end).toISOString(true), time_in: thetime.toISOString(true), approved: toapprove}).fetch();
+                } else {
+                    await Timesheet.create({name: record.name, unique: null, scheduled_in: null, scheduled_out: null, time_in: thetime.toISOString(true), approved: toapprove}).fetch();
+                }
 
                 // Update the director presence
                 await Directors.update({ID: record.ID}, {present: true, since: thetime.toISOString(true)})