+++ CHANGELOG.md
@@ -14,12 +14,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - Logging of when Google Calendar events were cancelled (removed) [only applicable for events within the next week to date]. Logged via "cancellation" logtype.
  - When logging an absence or cancellation, if the DJ does not exist in the system, it will be created instead of ignored.
  - sails.helpers.onesignal.send() and sails.helpers.onesignal.sendEvent() helpers for sending out push notifications.
+ - "scheduled_in", "scheduled_out", and "unique" attributes in Timesheet. This allows tracking office hour consistency.
+ - Logging when a director cancels office hours or does not clock in for office hours. These will fall under logtype "director-cancellation" and "director-absent" respectively.
 
 ### Changed
  - Many of the methods used in api controllers and in models have been migrated to sails helpers.
  - Removed a lot of the logging that could potentially output sensitive data to the console.
- - "calendar" model now uses nodebase datastore instead of RAM. This is so we can persist-store event cancellations (which would not be done through Google Calendar).
- - "active" parameter of calendar events will now be false either if the event passed the end time, or if the event was marked as canceled.
+ - "calendar" model now uses nodebase datastore instead of RAM. This is so we can persist-store event cancellations.
+ - "directorhours" model now uses timesheets datastore instead of RAM. This is so we can persist-store office hour cancellations.
  - requests/place response object changed. HTML property removed; instead, message property will be used, which will NOT contain div/bootstrap data.
  - sails.helpers.requests.checkRequestable now returns message property instead of HTML property; message property does not contain div/bootstrap data.
  - "active" property of calendar events is now a number instead of boolean: -1 is cancelled, 0 is expired, 1 is active.